Upd_11-10

+++ spring_projeto/src/main/java/com/entra21/grupo1/controller/PessoaRestController.java
@@ -39,7 +39,7 @@ public class PessoaRestController {
     }
 
     @PutMapping("/movimentacao")
-    public void addMovimentacao(@RequestBody Long valor) {
+    public void addMovimentacao(@RequestBody Double valor) {
         pessoaService.movimentacao(valor);
     }
